requirements.txt
# Ultra-lightweight requirements for Railway free tier (0.5GB RAM)
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Essential data processing (lightweight)
pandas==2.1.4
numpy==1.24.4

# Minimal ML (only sklearn for basic classification)
scikit-learn==1.3.2

# Web and networking
httpx==0.25.2
tldextract==5.1.1
dnspython==2.4.2

# Additional utilities
pydantic-settings==2.1.0